[Data Framework] Update candidate list to use data framework (#4918)
This updates the candidate_list module to use the LORIS data framework. It does so by adding a new abstract DataFrameworkMenu class, which can be extended in order to remove some of the boilerplate of converting menu filters to the data framework.

In order to use the DataFrameworkMenu class, an implementation must supply:

    a getFieldOptions function which returns a list of fieldOptions to include in the array for the frontend
    a getBaseDataProvisioner function which returns a data provisioner
    a useSiteFilter function which returns whether or not the data provisioner should have the UserSiteFilter match applied.

In return, the class takes care of using that dataprovisioner, adding default LORIS data filters, and serializing to JSON in the format expected by react menu filters.