[Imaging Browser] Fix menu loading (#6943)
The query was updated for the imaging browser to use an
AcquisitionDate column from the file table instead of a
join to the parameter_type table. However, a leftover
clause in the WHERE part of the SQL statement is referencing
the no longer joined table, causing the page to crash on
accessing the menu.

This removes the invalid clause.